Refactor completo en el modulo home se encuentran todo
