--- lib/barcode_photo/barcode_photo_injection.dart
@@ -1,17 +1,0 @@
-import 'package:flutter/material.dart';
-
-import 'package:provider/provider.dart';
-
-import 'package:qr_reader/barcode_photo/domain/barcode_reader_provider.dart';
-import 'package:qr_reader/barcode_photo/interface/barcode_reader_screen.dart';
-
-class BarcodePhotoInjection {
-  BarcodePhotoInjection._();
-
-  static Widget injection() {
-    return ListenableProvider(
-      create: (context) => BarcodePhotoProvider(),
-      child: const BarcodePhotoScreen(),
-    );
-  }
-}
