+++ lib/home/interface/home_screen.dart
@@ -1,13 +1,18 @@
 import 'package:flutter/material.dart';
+
+import 'package:provider/provider.dart';
+import 'package:flutter_barcode_scanner/flutter_barcode_scanner.dart';
 import 'package:google_fonts/google_fonts.dart';
-import 'package:qr_reader/barcode_photo/barcode_photo_injection.dart';
-import 'package:qr_reader/qr_reader/qr_reader_injection.dart';
+
+import 'package:qr_reader/home/domain/home_provider.dart';
+import 'package:url_launcher/url_launcher.dart';
 
 class HomeScreen extends StatelessWidget {
   const HomeScreen({super.key});
 
   @override
   Widget build(BuildContext context) {
+    final provider = Provider.of<HomeProvider>(context);
     return Scaffold(
       appBar: AppBar(
         backgroundColor: Colors.deepPurple,
@@ -25,62 +30,109 @@ class HomeScreen extends StatelessWidget {
         child: Column(
           mainAxisAlignment: MainAxisAlignment.center,
           children: [
-            NavigatorButton(
-              title: 'QR Scanner',
-              navegacion: QrReaderInjection.injection(),
+            InkWell(
+              onTap: () async {
+                String scannedCode = await FlutterBarcodeScanner.scanBarcode(
+                  '#ff6666',
+                  'Cancelar',
+                  true,
+                  ScanMode.QR,
+                );
+                if (scannedCode != '-1') {
+                  Uri url = Uri.parse(scannedCode);
+                  if (!await launchUrl(url)) {
+                    throw 'No se puede abrir la URL $scannedCode';
+                  }
+                }
+              },
+              borderRadius: BorderRadius.circular(16),
+              child: Container(
+                padding: const EdgeInsets.symmetric(horizontal: 5),
+                decoration: BoxDecoration(
+                  borderRadius: BorderRadius.circular(60),
+                  color: Colors.deepPurple,
+                ),
+                height: 35,
+                width: 155,
+                child: Center(
+                  child: Text(
+                    'QR Scanner',
+                    overflow: TextOverflow.ellipsis,
+                    maxLines: 1,
+                    style: GoogleFonts.nunito(),
+                  ),
+                ),
+              ),
             ),
             const SizedBox(
               height: 15,
             ),
-            NavigatorButton(
-              title: 'Foto del QR',
-              navegacion: BarcodePhotoInjection.injection(),
+            // BOTON FOTO PARA QR
+            InkWell(
+              onTap: () async {
+                await provider.pickImageQR();
+                if (provider.errorMessageQr != null) {
+                  ScaffoldMessenger.of(context).showSnackBar(
+                    SnackBar(
+                      content: Text(provider.errorMessageQr!),
+                    ),
+                  );
+                }
+              },
+              borderRadius: BorderRadius.circular(16),
+              child: Container(
+                padding: const EdgeInsets.symmetric(horizontal: 5),
+                decoration: BoxDecoration(
+                  borderRadius: BorderRadius.circular(60),
+                  color: Colors.deepPurple,
+                ),
+                height: 35,
+                width: 155,
+                child: Center(
+                  child: Text(
+                    'Foto QR',
+                    overflow: TextOverflow.ellipsis,
+                    maxLines: 1,
+                    style: GoogleFonts.nunito(),
+                  ),
+                ),
+              ),
             ),
             const SizedBox(
               height: 15,
             ),
-            NavigatorButton(
-              title: 'Foto código de barras',
-              navegacion: BarcodePhotoInjection.injection(),
+            // BOTON FOTO CODIGO DE BARRAS
+            InkWell(
+              onTap: () async {
+                await provider.pickImageBar();
+                if (provider.errorMessage != null) {
+                  ScaffoldMessenger.of(context).showSnackBar(
+                    SnackBar(
+                      content: Text(provider.errorMessage!),
+                    ),
+                  );
+                }
+              },
+              borderRadius: BorderRadius.circular(16),
+              child: Container(
+                padding: const EdgeInsets.symmetric(horizontal: 5),
+                decoration: BoxDecoration(
+                  borderRadius: BorderRadius.circular(60),
+                  color: Colors.deepPurple,
+                ),
+                height: 35,
+                width: 155,
+                child: Center(
+                  child: Text(
+                    'Foto código de barras',
+                    overflow: TextOverflow.ellipsis,
+                    maxLines: 1,
+                    style: GoogleFonts.nunito(),
+                  ),
+                ),
+              ),
             ),
           ],
-        ),
-      ),
-    );
-  }
-}
-
-class NavigatorButton extends StatelessWidget {
-  const NavigatorButton({super.key, required this.title, this.navegacion});
-  final String title;
-  final dynamic navegacion;
-  @override
-  Widget build(BuildContext context) {
-    return InkWell(
-      onTap: () {
-        Navigator.push(
-          context,
-          MaterialPageRoute(
-            builder: (context) => navegacion,
-          ),
-        );
-      },
-      borderRadius: BorderRadius.circular(16),
-      child: Container(
-        padding: const EdgeInsets.symmetric(horizontal: 5),
-        decoration: BoxDecoration(
-          borderRadius: BorderRadius.circular(60),
-          color: Colors.deepPurple,
-        ),
-        height: 35,
-        width: 155,
-        child: Center(
-          child: Text(
-            title,
-            overflow: TextOverflow.ellipsis,
-            maxLines: 1,
-            style: GoogleFonts.nunito(),
-          ),
         ),
       ),
     );
